requirements.txt
pyqt5>=5.9.2
pyopengl
numpy
Pillow
PlyPlus